+++ include/swift/Strings.h
@@ -26,6 +26,8 @@ constexpr static const StringLiteral SWIFT_ONONE_SUPPORT = "SwiftOnoneSupport";
 constexpr static const StringLiteral SWIFT_SHIMS_NAME = "SwiftShims";
 /// The name of the Builtin module, which contains Builtin functions.
 constexpr static const StringLiteral BUILTIN_NAME = "Builtin";
+/// The name of the clang imported header module.
+constexpr static const StringLiteral CLANG_HEADER_MODULE_NAME = "__ObjC";
 /// The prefix of module names used by LLDB to capture Swift expressions
 constexpr static const StringLiteral LLDB_EXPRESSIONS_MODULE_NAME_PREFIX =
     "__lldb_expr_";
