+++ lib/Serialization/DeserializationErrors.h
@@ -21,8 +21,10 @@
 
 namespace swift {
 class ModuleFile;
+class ModuleFileSharedCore;
 
 StringRef getNameOfModule(const ModuleFile *);
+StringRef getNameOfModule(const ModuleFileSharedCore *);
 
 namespace serialization {
 
@@ -453,6 +455,17 @@ public:
 
   void print(raw_ostream &os) const override {
     os << Action << " \'" << getNameOfModule(&MF) << "'\n";
+  }
+};
+
+class PrettyStackTraceModuleFileCore : public llvm::PrettyStackTraceEntry {
+  const ModuleFileSharedCore &MF;
+public:
+  explicit PrettyStackTraceModuleFileCore(ModuleFileSharedCore &module)
+      : MF(module) {}
+
+  void print(raw_ostream &os) const override {
+    os << "While reading from \'" << getNameOfModule(&MF) << "'\n";
   }
 };
